+++ NEWS.md
@@ -2,6 +2,7 @@
 * Fix for Philips svs spar/sdat voxel geometry.
 * Bug fixes for match_lineshape function.
 * Added zero_spec_threshold function.
+* Added set_spec_values function.
 
 # spant 4.3.0
 * Bug fix for mr_data2bids function.